Category list with RecyclerView.

+++ app/src/main/java/br/fef/data/persistence/entity/User.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "user")
 data class User(
-    @PrimaryKey var cod: Long,
+    @PrimaryKey var cod: Int,
     var nome: String,
     var email: String,
     var tipo: Int
